Set up Makefile for example
I'm assuming that the bmi-fortran example has been installed in
`../_install`. Point to its `lib` directory with the `-L` flag.
Point to Fortran `.mod` files with the `-I` flag. Set the rpath
to this nonstandard library location with `-Wl,rpath`. Explicitly
link against the *bmif* and *bmiheatf* shared libraries in this
location.