Extend example by adding pointer to derived type
This is where it gets interesting.

I have a Fortran pointer in the type defined in the pure Fortran
module. In the interoperable Fortran module, I shadow the pointer in
the type with a `c_ptr`. With this, I can exchange information between
the pure Fortran module and a C program.